security

+++ src/main/java/dev/dewy/dqs/discord/dewy/KillCommand.java
@@ -27,57 +27,60 @@ public class KillCommand extends Command
     @Override
     protected void execute(CommandEvent event)
     {
-        try
+        if (event.getAuthor().getId().equals(Constants.CONFIG.discord.operatorId) && event.getChannel().getId().equals(Constants.CONFIG.discord.channelId))
         {
-            if (event.getAuthor().getId().equals(Constants.CONFIG.discord.operatorId))
+            try
             {
+                if (event.getAuthor().getId().equals(Constants.CONFIG.discord.operatorId))
+                {
+                    event.reply(new EmbedBuilder()
+                            .setTitle("**DQS** - Instance Termination")
+                            .setDescription("Shutting down instance for " + event.getJDA().getUserById(Constants.CONFIG.discord.subscriberId).getName() + "...")
+                            .setColor(new Color(10144497))
+                            .setFooter("Focused on " + Constants.CONFIG.authentication.username, new URL(String.format("https://crafatar.com/avatars/%s?size=64&overlay&default=MHF_Steve", Constants.CONFIG.authentication.uuid)).toString())
+                            .setAuthor("DQS " + Constants.VERSION, null, "https://i.imgur.com/xTd3Ri3.png")
+                            .build());
+
+                    Thread.sleep(10000L);
+
+                    if (DQS.getInstance().server != null)
+                    {
+                        DQS.getInstance().server.close(true);
+                    }
+
+                    WEBSOCKET_SERVER.shutdown();
+                    saveConfig();
+
+                    Runtime.getRuntime().exit(0);
+                }
+
                 event.reply(new EmbedBuilder()
-                        .setTitle("**DQS** - Instance Termination")
-                        .setDescription("Shutting down instance for " + event.getJDA().getUserById(Constants.CONFIG.discord.subscriberId).getName() + "...")
-                        .setColor(new Color(10144497))
-                        .setFooter("Focused on " + Constants.CONFIG.authentication.username, new URL(String.format("https://crafatar.com/avatars/%s?size=64&overlay&default=MHF_Steve", Constants.CONFIG.authentication.uuid)).toString())
+                        .setTitle("**DQS** - Insufficient Permissions")
+                        .setDescription("This command can only be executed by an operator.")
+                        .setColor(new Color(15221016))
+                        .setFooter("Focused on " + Constants.CONFIG.authentication.username)
+                        .setAuthor("DQS " + Constants.VERSION, null, "https://i.imgur.com/xTd3Ri3.png")
+                        .build());
+            } catch (Throwable t)
+            {
+                Constants.DISCORD_LOG.error(t);
+
+                event.reply(new EmbedBuilder()
+                        .setTitle("**DQS** - Error")
+                        .setDescription("An exception occurred whilst executing this command. Debug information has been sent to Dewy to be fixed in following updates. Sorry about any inconvenience!")
+                        .setColor(new Color(15221016))
+                        .setFooter("Focused on " + Constants.CONFIG.authentication.username)
                         .setAuthor("DQS " + Constants.VERSION, null, "https://i.imgur.com/xTd3Ri3.png")
                         .build());
 
-                Thread.sleep(10000L);
-
-                if (DQS.getInstance().server != null)
-                {
-                    DQS.getInstance().server.close(true);
-                }
-
-                WEBSOCKET_SERVER.shutdown();
-                saveConfig();
-
-                Runtime.getRuntime().exit(0);
+                Objects.requireNonNull(event.getJDA().getUserById(Constants.CONFIG.discord.operatorId)).openPrivateChannel().queue((privateChannel ->
+                        privateChannel.sendMessage(new EmbedBuilder()
+                                .setTitle("**DQS** - Error Report (" + Objects.requireNonNull(event.getJDA().getUserById(Constants.CONFIG.discord.subscriberId)).getName() + ")")
+                                .setDescription("A " + t.getClass().getSimpleName() + " was thrown during the execution of a kill command.\n\n**Cause:**\n\n```" + t.getMessage() + "```")
+                                .setColor(new Color(15221016))
+                                .setAuthor("DQS " + Constants.VERSION, null, "https://i.imgur.com/xTd3Ri3.png")
+                                .build()).queue()));
             }
-
-            event.reply(new EmbedBuilder()
-                    .setTitle("**DQS** - Insufficient Permissions")
-                    .setDescription("This command can only be executed by an operator.")
-                    .setColor(new Color(15221016))
-                    .setFooter("Focused on " + Constants.CONFIG.authentication.username)
-                    .setAuthor("DQS " + Constants.VERSION, null, "https://i.imgur.com/xTd3Ri3.png")
-                    .build());
-        } catch (Throwable t)
-        {
-            Constants.DISCORD_LOG.error(t);
-
-            event.reply(new EmbedBuilder()
-                    .setTitle("**DQS** - Error")
-                    .setDescription("An exception occurred whilst executing this command. Debug information has been sent to Dewy to be fixed in following updates. Sorry about any inconvenience!")
-                    .setColor(new Color(15221016))
-                    .setFooter("Focused on " + Constants.CONFIG.authentication.username)
-                    .setAuthor("DQS " + Constants.VERSION, null, "https://i.imgur.com/xTd3Ri3.png")
-                    .build());
-
-            Objects.requireNonNull(event.getJDA().getUserById(Constants.CONFIG.discord.operatorId)).openPrivateChannel().queue((privateChannel ->
-                    privateChannel.sendMessage(new EmbedBuilder()
-                            .setTitle("**DQS** - Error Report (" + Objects.requireNonNull(event.getJDA().getUserById(Constants.CONFIG.discord.subscriberId)).getName() + ")")
-                            .setDescription("A " + t.getClass().getSimpleName() + " was thrown during the execution of a kill command.\n\n**Cause:**\n\n```" + t.getMessage() + "```")
-                            .setColor(new Color(15221016))
-                            .setAuthor("DQS " + Constants.VERSION, null, "https://i.imgur.com/xTd3Ri3.png")
-                            .build()).queue()));
         }
     }
 }
